Rename the module for the list section handler wrapper
The organization of utils in ingredient-handlers is a mess. I'll need to
fix this in a follow up PR.
