+++ scripts/scraper-ng/rules/html-require-recipe-ingredients/ingredient-handlers/list-sections.js
